refactored huge management sections tests file; updated view permissions not only for unicef where possible; reverted partner intervention serializer to use shared class; migrated intervention cash transfer modalities to be list field
