+++ src/etools/applications/partners/tests/_api_checker/etools.applications.partners.tests.test_api/TestAPIIntervention/_api_v2_interventions_/get/None.response.json
@@ -80,7 +80,7 @@
             "ip_program_contribution": "",
             "budget_owner": null,
             "hq_support_cost": "0.0",
-            "cash_transfer_modalities": "dct",
+            "cash_transfer_modalities": ["dct"],
             "unicef_review_type": "none"
         },
         {
