+++ src/etools/applications/partners/tests/_api_checker/etools.applications.partners.tests.test_api/TestAPIIntervention/_api_v2_interventions_101_/get/None.response.json
@@ -19,7 +19,7 @@
         ],
         "content-length": [
             "Content-Length",
-            "7729"
+            "7818"
         ]
     },
     "data": {
@@ -110,7 +110,7 @@
                 "status": true,
                 "sections": false,
                 "modified": true,
-                "risks": true,
+                "risks": false,
                 "unicef_review_type": true,
                 "actionpoint": true,
                 "technical_guidance": true,
@@ -161,6 +161,7 @@
                 "offices": false,
                 "hq_support_cost": true,
                 "sustainability_narrative": true,
+                "final_partnership_review": false,
                 "frs": false,
                 "signed_by_partner_date": false,
                 "partner_focal_points": false,
@@ -177,10 +178,10 @@
                 "start": false,
                 "country_programme": false,
                 "prc_review_document": false,
-                "cash_transfer_modalities": true,
+                "cash_transfer_modalities": false,
                 "travel_activities": true,
                 "sections_present": true,
-                "pd_outputs": true
+                "pd_outputs": false
             },
             "required": {
                 "signed_pd_document": false,
@@ -247,6 +248,7 @@
                 "hq_support_cost": false,
                 "sustainability_narrative": false,
                 "frs": false,
+                "final_partnership_review": false,
                 "signed_by_partner_date": false,
                 "partner_focal_points": false,
                 "monitoring_activities": false,
@@ -272,10 +274,10 @@
                 "reporting_requirements": false,
                 "agreement": false,
                 "unicef_court": true,
-                "planned_visits": false,
+                "planned_visits": true,
                 "cfei_number": true,
                 "sustainability_rating": true,
-                "unicef_signatory": false,
+                "unicef_signatory": true,
                 "unicef_signatory_id": true,
                 "status": true,
                 "sections": false,
@@ -284,15 +286,15 @@
                 "unicef_review_type": true,
                 "actionpoint": true,
                 "technical_guidance": true,
-                "review_date_prc": false,
+                "review_date_prc": true,
                 "reporting_periods": true,
                 "management_budgets": true,
                 "supply_items": true,
                 "context": true,
                 "gender_rating": true,
-                "submission_date": false,
-                "partner_authorized_officer_signatory": false,
-                "amendments": false,
+                "submission_date": true,
+                "partner_authorized_officer_signatory": true,
+                "amendments": true,
                 "end": false,
                 "budget_owner": true,
                 "special_reporting_requirements": true,
@@ -301,16 +303,16 @@
                 "number": false,
                 "created": true,
                 "title": false,
-                "submission_date_prc": false,
+                "submission_date_prc": true,
                 "termination_doc": false,
-                "signed_pd_attachment": false,
+                "signed_pd_attachment": true,
                 "agreement_id": true,
                 "equity_narrative": true,
                 "activation_letter_attachment": false,
                 "result_links": true,
                 "attachments": true,
                 "humanitarian_flag": true,
-                "signed_by_unicef_date": false,
+                "signed_by_unicef_date": true,
                 "document_type": false,
                 "unicef_focal_points": false,
                 "partner_authorized_officer_signatory_id": true,
@@ -331,11 +333,12 @@
                 "offices": false,
                 "hq_support_cost": true,
                 "sustainability_narrative": true,
-                "frs": false,
-                "signed_by_partner_date": false,
+                "final_partnership_review": true,
+                "frs": true,
+                "signed_by_partner_date": true,
                 "partner_focal_points": false,
                 "monitoring_activities": true,
-                "prc_review_attachment": false,
+                "prc_review_attachment": true,
                 "id": true,
                 "planned_budget": false,
                 "metadata": true,
