+++ src/etools/applications/partners/tests/test_epd_management/final_partnership_review.py
@@ -1,0 +1,92 @@
+from django.core.files.uploadedfile import SimpleUploadedFile
+from django.urls import reverse
+
+from rest_framework import status
+
+from etools.applications.attachments.tests.factories import AttachmentFactory
+from etools.applications.partners.models import FileType
+from etools.applications.partners.tests.factories import FileTypeFactory
+from etools.applications.partners.tests.test_epd_management.base import BaseTestCase
+
+
+class TestFinalPartnershipReviewManagement(BaseTestCase):
+    @classmethod
+    def setUpTestData(cls):
+        super().setUpTestData()
+        FileTypeFactory(name=FileType.FINAL_PARTNERSHIP_REVIEW)
+        cls.example_attachment = AttachmentFactory(
+            file=SimpleUploadedFile('hello_world.txt', b'hello world!'),
+        )
+
+    # test permissions
+    def test_unicef_permissions(self):
+        response = self.forced_auth_req(
+            'get',
+            reverse('pmp_v3:intervention-detail', args=[self.ended_intervention.pk]),
+            user=self.unicef_user,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.data)
+        self.assertEqual(response.data['permissions']['view']['final_partnership_review'], True)
+        self.assertEqual(response.data['permissions']['edit']['final_partnership_review'], False)
+
+    def test_partnership_manager_permissions(self):
+        response = self.forced_auth_req(
+            'get',
+            reverse('pmp_v3:intervention-detail', args=[self.draft_intervention.pk]),
+            user=self.partnership_manager,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.data)
+        self.assertEqual(response.data['permissions']['view']['final_partnership_review'], True)
+        self.assertEqual(response.data['permissions']['edit']['final_partnership_review'], False)
+
+    def test_partnership_manager_permissions_ended(self):
+        response = self.forced_auth_req(
+            'get',
+            reverse('pmp_v3:intervention-detail', args=[self.ended_intervention.pk]),
+            user=self.partnership_manager,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.data)
+        self.assertEqual(response.data['permissions']['view']['final_partnership_review'], True)
+        self.assertEqual(response.data['permissions']['edit']['final_partnership_review'], True)
+
+    def test_partner_permissions(self):
+        response = self.forced_auth_req(
+            'get',
+            reverse('pmp_v3:intervention-detail', args=[self.ended_intervention.pk]),
+            user=self.partner_focal_point,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.data)
+        self.assertEqual(response.data['permissions']['view']['final_partnership_review'], False)
+        self.assertEqual(response.data['permissions']['edit']['final_partnership_review'], False)
+
+    # test functionality
+    def test_update(self):
+        response = self.forced_auth_req(
+            'patch',
+            reverse('pmp_v3:intervention-detail', args=[self.ended_intervention.pk]),
+            user=self.partnership_manager,
+            data={
+                'final_partnership_review': self.example_attachment.id,
+            }
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK, response.data)
+        self.assertEqual(response.data['permissions']['edit']['final_partnership_review'], True)
+        self.assertIn('hello_world.txt', response.data['final_partnership_review']['attachment'])
+        self.assertEqual(
+            self.ended_intervention.attachments.get(
+                type__name=FileType.FINAL_PARTNERSHIP_REVIEW
+            ).attachment_file.get().pk,
+            self.example_attachment.pk
+        )
+
+    def test_update_permissions_restricted(self):
+        response = self.forced_auth_req(
+            'patch',
+            reverse('pmp_v3:intervention-detail', args=[self.draft_intervention.pk]),
+            user=self.partnership_manager,
+            data={
+                'final_partnership_review': self.example_attachment.pk,
+            }
+        )
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertEqual(response.data[0], 'Cannot change fields while in draft: final_partnership_review')
