+++ src/etools/applications/partners/views/interventions_v3.py
@@ -22,7 +22,7 @@ from etools.applications.partners.models import (
 )
 from etools.applications.partners.permissions import (
     intervention_field_is_editable_permission,
-    PartnershipManagerPermission,
+    PMPInterventionPermission,
 )
 from etools.applications.partners.serializers.exports.interventions import (
     InterventionExportFlatSerializer,
@@ -38,7 +38,7 @@ from etools.applications.partners.serializers.interventions_v3 import (
     InterventionDummySerializer,
     InterventionManagementBudgetSerializer,
     InterventionRiskSerializer,
-    InterventionSupplyItemSerializer, PartnerInterventionDetailSerializer,
+    InterventionSupplyItemSerializer,
 )
 from etools.applications.partners.serializers.v3 import (
     PartnerInterventionLowerResultSerializer,
@@ -50,12 +50,50 @@ from etools.applications.reports.models import InterventionActivity, LowerResult
 from etools.applications.reports.serializers.v2 import InterventionActivityDetailSerializer
 
 
+class APIActionsMixin:
+    """
+    add viewsets-like action attribute to generic api views to reuse action-based things, for example permissions
+    """
+    action = None
+    detail = None
+
+    def get_action(self, method):
+        if method == 'OPTIONS':
+            return 'metadata'
+
+        if not self.detail:
+            if method == 'GET':
+                return 'list'
+            elif method == 'POST':
+                return 'create'
+        else:
+            if method == 'GET':
+                return 'retrieve'
+            elif method == 'PUT':
+                return 'update'
+            elif method == 'PATCH':
+                return 'partial_update'
+            elif method == 'DELETE':
+                return 'delete'
+
+        return 'unknown'
+
+    def dispatch(self, request, *args, **kwargs):
+        # if api view is inherited from one of GenericAPIView subclasses, we can just check which methods are defined
+        if hasattr(self, 'list') or hasattr(self, 'create'):
+            self.detail = False
+        else:
+            self.detail = True
+        self.action = self.get_action(request.method.upper())
+        return super().dispatch(request, *args, **kwargs)
+
+
 class PMPInterventionMixin(PMPBaseViewMixin):
     SERIALIZER_OPTIONS = {
-        "list": (InterventionListSerializer, InterventionDummySerializer),
-        "create": (InterventionCreateUpdateSerializer, InterventionDummySerializer),
-        "detail": (InterventionDetailSerializer, PartnerInterventionDetailSerializer),
-        "list_min": (MinimalInterventionListSerializer, InterventionDummySerializer),
+        "list": (InterventionListSerializer, InterventionListSerializer),
+        "create": (InterventionCreateUpdateSerializer, InterventionCreateUpdateSerializer),
+        "detail": (InterventionDetailSerializer, InterventionDetailSerializer),
+        "list_min": (MinimalInterventionListSerializer, MinimalInterventionListSerializer),
         "csv": (InterventionExportSerializer, InterventionDummySerializer),
         "csv_flat": (InterventionExportFlatSerializer, InterventionDummySerializer),
     }
@@ -68,8 +106,8 @@ class PMPInterventionMixin(PMPBaseViewMixin):
         return qs
 
 
-class PMPInterventionListCreateView(PMPInterventionMixin, InterventionListAPIView):
-    permission_classes = (IsAuthenticated, PartnershipManagerPermission,)
+class PMPInterventionListCreateView(APIActionsMixin, PMPInterventionMixin, InterventionListAPIView):
+    permission_classes = (IsAuthenticated, PMPInterventionPermission)
 
     def get_serializer_class(self):
         if self.request.method == "GET":
@@ -97,7 +135,7 @@ class PMPInterventionListCreateView(PMPInterventionMixin, InterventionListAPIVie
         )
 
 
-class PMPInterventionRetrieveUpdateView(PMPInterventionMixin, InterventionDetailAPIView):
+class PMPInterventionRetrieveUpdateView(APIActionsMixin, PMPInterventionMixin, InterventionDetailAPIView):
     SERIALIZER_MAP = copy(InterventionDetailAPIView.SERIALIZER_MAP)
     SERIALIZER_MAP['risks'] = InterventionRiskSerializer
     related_fields = InterventionDetailAPIView.related_fields + [
